# api/wellness-buddy-openapi.yaml
openapi: 3.0.3
info:
  title: Wellness Buddy API
  version: 1.0.0
  description: |
    REST API for the Wellness Buddy app. The client must never access the
    database directly; all reads and writes go through this API layer.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /expenses:
    get:
      tags: [Expenses]
      summary: List expenses
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
    post:
      tags: [Expenses]
      summary: Create an expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
  /expenses/{expenseId}:
    get:
      tags: [Expenses]
      summary: Get an expense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '200':
          description: Expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Expenses]
      summary: Update an expense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseUpdate'
      responses:
        '200':
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Expenses]
      summary: Delete an expense
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '204':
          description: Expense deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /clubs:
    get:
      tags: [Clubs]
      summary: List clubs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
    post:
      tags: [Clubs]
      summary: Create a club
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubCreate'
      responses:
        '201':
          description: Club created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
  /clubs/{clubId}:
    get:
      tags: [Clubs]
      summary: Get a club
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClubId'
      responses:
        '200':
          description: Club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Clubs]
      summary: Update a club
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClubId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubUpdate'
      responses:
        '200':
          description: Updated club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Clubs]
      summary: Delete a club
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClubId'
      responses:
        '204':
          description: Club deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /contract-templates:
    get:
      tags: [Legal Templates]
      summary: List contract templates
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContractTemplate'
    post:
      tags: [Legal Templates]
      summary: Create a contract template
      description: For moderator/admin use; normal users only read.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTemplateCreate'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
  /contract-templates/{templateId}:
    get:
      tags: [Legal Templates]
      summary: Get a template
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Legal Templates]
      summary: Update a template
      description: Typically restricted to moderators/admins.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractTemplateUpdate'
      responses:
        '200':
          description: Updated template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Legal Templates]
      summary: Delete a template
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '204':
          description: Template deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /ai/recommendations:
    get:
      tags: [AI Recommendations]
      summary: List AI recommendations
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIRecommendation'
    post:
      tags: [AI Recommendations]
      summary: Create an AI recommendation
      description: For backend/cron jobs; not exposed in client UI.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRecommendationCreate'
      responses:
        '201':
          description: Recommendation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRecommendation'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ExpenseId:
      name: expenseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ClubId:
      name: clubId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TemplateId:
      name: templateId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Expense:
      type: object
      properties:
        expense_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
        amount:
          type: number
          format: float
        date:
          type: string
          format: date
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExpenseCreate:
      type: object
      required: [user_id, category, amount, date]
      properties:
        user_id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
        amount:
          type: number
          format: float
        date:
          type: string
          format: date
        description:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        goal_id:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
        amount:
          type: number
          format: float
        date:
          type: string
          format: date
        description:
          type: string
    Club:
      type: object
      properties:
        club_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        created_at:
          type: string
          format: date-time
    ClubCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
    ClubUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
...
    ContractTemplate:
      type: object
      properties:
        template_id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        category:
          type: string
        description:
          type: string
        file_url:
          type: string
        created_at:
          type: string
          format: date-time
    ContractTemplateCreate:
      type: object
      required: [title, file_url]
      properties:
        creator_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        category:
          type: string
        description:
          type: string
        file_url:
          type: string
    ContractTemplateUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        file_url:
          type: string
...
    AIRecommendation:
      type: object
      properties:
        recommendation_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        category:
          type: string
        message:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
    AIRecommendationCreate:
      type: object
      required: [user_id, message]
      properties:
        user_id:
          type: string
          format: uuid
        category:
          type: string
        message:
          type: string
        source:
          type: string
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true